move Monitoring to evolution-frontend
